requirements.txt
# ====================================
# DBT + SNOWFLAKE
# ====================================
dbt-core==1.11.2
dbt-snowflake==1.11.1
snowflake-connector-python==4.2.0


# ====================================
# AWS 
# ====================================
boto3==1.42.34

# ====================================
# UTILITIES
# ====================================
python-dotenv==1.2.1

# ====================================
# AIRFLOW PROVIDERS
# ====================================
apache-airflow-providers-snowflake==5.8.1